Add max paths check to validation script
