+++ _scripts/printAllSquaresDistrictId.py
@@ -1,4 +1,4 @@
-import frb
+import _scripts.validation.frb as frb
 from pathlib import Path
 from termcolor import cprint
 import colorama
