+++ _scripts/validation/naming.py
@@ -3,7 +3,7 @@ from validation.errors import process_strErrors
 
 def check_naming_convention(name, yamlMap):
     strErrors = []
-    print(f"{name:24} Naming Convention Check...", end="")
+    print(f"{name:24} Naming Convention Check......", end="")
 
     if " " in yamlMap.name:
         strErrors.append(f"There is a whitespace character in {yamlMap}")
